PatchMemory, WriteNop/Jmp -> PatchNop/Jmp, Unpatch
Patch* functions store the original bytes in an unordered map, later to be reverted in Unpatch. PatchMemory is a wrapper for WriteMemory with the aforementioned functionality. This build fixes leftover JMP hooks crashing the game, as hooks::Destroy() now properly gets rid of them. This build also unpatches everything under 'patches.cpp'.