turn a memcard related patch into a hook
loading/creating a save will return you to the menu without reloading the button layout, causing various issues (for example: if you were already loaded in a save with "resume career", and you create a new save, the "resume career" button will persist, which it shouldn't (new saves can only start career, not resume it), and causes weird things to happen). in FECareerStateManager::HandleChildFlowDone, for this->mCurState == 15, it calls resume career, which was patched out. this patch has now been removed as it has been replaced with a hook which calls FEStateManager::Switch (exactly what FECareerStateManager::ResumeCareer uses when the tutorial hasn't been completed and career hasn't been started on the new save) which reloads the menu and thus the button layout. as both creating saves (previously 14) and creating saves use state 15, this effectively reloads the button layout in both cases and eliminates the bug entirely